Bug Fix - dyno client can hang client application on startup
If the host supplier reports hosts are available but dyno cannot establish connections with them on startup exceptions are thrown for each failed connection. If each exception stack trace is being logged the java process spikes the cpu at 100%. Also the connection pool itself is left in an unrecoverable state since the pool is marked as down but it's in the host mapping.