Clearly distinguish between operation failures and failover counts
Modified the behavior such that if the client is configured to use failover then a failure is counted after attempts are made at finding both local and failover connection pools to borrow a connection from, i.e. if the local selector has no hosts available but a failover is found and succeeds then an operation failure is not counted. Currently every failover attempt is also recorded as an operation failure regardless of whether failover succeeds.