Bug fix in host metrics count
The counter class relied on synchronizing it's collection with that of the host supplier via add/remove hosts however this can get out of sync, for example, if the client cluster starts while dynomite server(s) are down. In this case the host count is off as is the number of hosts down.