Bug Fix - host pool can run out of connections
Detailed Description
The 1.4.0 release discards connections that encounter socket exceptions due to an insidious underlying buffer issue. If the dynomite connection is severed while all connections are in use all of them will be discarded and the pool will not replenish itself immediately. The result is that all requests will fail-over to a secondary AZ. The fail-over is by design and the correct behavior and if the HostSupplier notifies the client that dynomite is Down, all is well. If the network connection simply had an issue and Dynomite is actually UP, then the HostSupplier will not notify the client and the pool will never replenish its connections to that host.

This is particularly problematic if the client application has wrapped the dyno client call in a load-shedding layer such as Hystrix and left the default timeouts in place (Hystrix is 1 second, Dyno client timeout default is 2 seconds) therefore since the client app timeout fires first the fallback for dyno client does not occur and all requests fail.

Solution
Detect when the pool is exhausted and notify the health tracker which will immediately put the pool into a reconnect state after which, all of the client's current reconnect + failover logic will apply.