Cross-zone (rack) fallback implementation
Currently cross-zone fallback only occurs when dyno client recognizes that a dynomite server is down. With this change any time an error occurs either obtaining a connection from the pool or with the execution of an operation the connection for the retry will be borrowed from a remote zone rather than the same local host pool where the error just occurred.

This change also contains the following minor fixes:
* Bug fix in failover metric reporting
* Automatic disabling of local zone affinity if the local zone has not been set. This often happens w/local laptop development + unit testing.
* Cleanup of 'datacenter' terminology to rack/zone. Initially datacenter was equivalent to zone/rack in Dynomite terminology. This changed however the dyno client code was never changed to reflect this which makes reading the code confusing.