Reduce calls to TokenMapSupplier
At the startup, the client was talking to every node in the dynomite cluster to get token information.
This was done at 3 places. This lead to slow client startup and also too many calls to the underlying token database,
This is a simple optimization to skip talking to remaining hosts if the client received the information for all the hosts
specfied.