Disable demo properties that affect the client in production
Somehow, the demo properties get picked up by the client even though
it's not supposed to which causes incorrect behavior in the client.

The properties in question are LOCAL_RACK and LOCAL_DATACENTER which
return the values set in the demo.properties file in production, causing
ConfigUtils.getLocalZone() and ConfigUtils.getDataCenter() to return
potentially incorrect values.