Dual writer in the DynoJedisDemo both point to primary cluster
Turns out that the 'shadowClusterClient' in DynoJedisDemo was pointing
to the primary cluster because it was fed the wrong HostSupplier and
TokenMapSupplier.

This was never noticed since the shadowClusterClient would verify if all
the data was there in the source cluster instead of the shadow cluster,
hence making all the dual writer based runs false positives.