Allow client to set per-connection level consistency
The Dynomite server side exposed an unsupported command called
"DYNO_CONFIG:CONN_CONSISTENCY <consistency_level>". This patch exposes
a new API on the DynoJedisClientBuilder that allows us to set
a consistency level per connection pool by invoking:

    withConnectionPoolConsistency( consistency )

If this API is invoked, on every connection to the Dynomite server, the
first command run will be the above mentioned one to set the consistency
level for that connection based on the user-provided arugument.

Testing: Used the DynoJedisDemo to try to get a key with quorum failures
and made sure it succeeds if withConnectionPoolConsistency("DC_ONE") is
set.

TODO: This was done in a hurry. Consider code-cleanup in the future.